refactor(chat): give tool-part interpretation its own module
